// api/mock-server/src/generated/mock-admin/contract/types.gen.ts
export type DeckRecord = DomainsDecksDeck & {
    id?: string;
    deletedAt?: string;
};
// ...
export type VisualIconName = string;

export type Id = string;

export type DateTime = string;

export type ActiveWorkspace = DomainsWorkspacesActiveWorkspace;

export type DomainsWorkspacesActiveWorkspace = {
    workspaceId: Id;
};

export type Folder = DomainsFoldersFolder;

export type DomainsFoldersFolder = {
    description: string;
    id: Id;
    name: string;
    parentId: Id;
    updatedAt: DateTime;
    workspaceId: Id;
};

export type Deck = DomainsDecksDeck;

export type DomainsDecksDeck = {
    description: string;
    dueToday: number;
    icon: VisualIconName;
    id: Id;
    parentId: Id;
    progress: number;
    title: string;
    totalNotes: number;
    updatedAt: DateTime;
    workspaceId: Id;
};